# swagger.ua.yaml
openapi: 3.0.0
info:
  title: Your Nexus Archive API
  description: |-
    API для мобільного застосунку "Your Nexus Archive" для колекціонерів марок.
    Цей API дозволяє користувачам керувати своєю колекцією, списком бажаних марок та марок для обміну.
    Додано нові адміністративні ендпоінти для керування каталогом марок.
  version: 1.2.0
  license:
    name: Mozilla Public License 2.0 (MPL)
    url: https://www.mozilla.org/en-US/MPL/2.0/
servers:
  - url: http://localhost:4484
    description: Локальний сервер для розробки
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          example: 'Олександр'
        email:
          type: string
          example: 'oleksandr@example.com'
        password:
          type: string
          format: password
          example: 'Password123'
      required:
        - name
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          example: 'oleksandr@example.com'
        password:
          type: string
          format: password
          example: 'Password123'
      required:
        - email
        - password
    User:
      type: object
      properties:
        _id:
          type: string
          example: '66a88b5e9f8a371c646b9a84'
        name:
          type: string
          example: 'Олександр'
        email:
          type: string
          example: 'oleksandr@example.com'
        role:
          type: string
          enum: ['user', 'admin']
          example: 'user'
        collectedStamps:
          type: array
          items:
            type: string
            example: '66a88b5e9f8a371c646b9a85'
        desiredStamps:
          type: array
          items:
            type: string
            example: '66a88b5e9f8a371c646b9a86'
        stampsForExchange:
          type: array
          items:
            type: string
            example: '66a88b5e9f8a371c646b9a87'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        message:
          type: string
          example: 'Вхід успішний'
        data:
          type: object
          properties:
            accessToken:
              type: string
              example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY2YTg4YjVlOWY4YTM3MWM2NDZiOWE4NCIsImlhdCI6MTcyMjM3Mzc5MywiZXhwIjoxNzIyMzc3MzkzfQ.5m8m3R8l5X9q8b8W7n9N1z1m6M3V2d3E4'
            refreshToken:
              type: string
              example: '66a88b5e9f8a371c646b9a84'
            user:
              $ref: '#/components/schemas/User'
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: 'Успішно виконано'
        data:
          type: object
          description: Дані можуть відрізнятися залежно від запиту
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 404
        message:
          type: string
          example: 'Не знайдено'
        data:
          type: object
          properties:
            message:
              type: string
              example: 'Ресурс не знайдено.'
...
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Зареєструвати нового користувача
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Успішна реєстрація, повертає дані користувача
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Некоректні дані в запиті. Тіло запиту містить недійсні або відсутні дані.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Користувач з такою електронною поштою вже існує.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Вхід користувача
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успішний вхід, повертає токени та дані користувача.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неправильна електронна пошта або пароль.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: Вихід користувача
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Успішний вихід.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Несанкціонований доступ. Користувач не аутентифікований.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...